presentation-summarizer/requirements.txt
python-pptx==0.6.21
openai==1.3.0
python-dotenv==1.0.0
click==8.1.3
flask==2.3.3
werkzeug==2.3.7
